// registry-server/prisma/generated/zod/outputTypeSchemas/RoomUpsertArgsSchema.ts
import { z } from 'zod';
import type { Prisma } from '../../../../lib/generated/prisma';
import { RoomIncludeSchema } from '../inputTypeSchemas/RoomIncludeSchema'
import { RoomWhereUniqueInputSchema } from '../inputTypeSchemas/RoomWhereUniqueInputSchema'
import { RoomCreateInputSchema } from '../inputTypeSchemas/RoomCreateInputSchema'
import { RoomUncheckedCreateInputSchema } from '../inputTypeSchemas/RoomUncheckedCreateInputSchema'
import { RoomUpdateInputSchema } from '../inputTypeSchemas/RoomUpdateInputSchema'
import { RoomUncheckedUpdateInputSchema } from '../inputTypeSchemas/RoomUncheckedUpdateInputSchema'
import { FurnitureFindManyArgsSchema } from "../outputTypeSchemas/FurnitureFindManyArgsSchema"
import { CompanionFindManyArgsSchema } from "../outputTypeSchemas/CompanionFindManyArgsSchema"
import { RoomCountOutputTypeArgsSchema } from "../outputTypeSchemas/RoomCountOutputTypeArgsSchema"
// Select schema needs to be in file to prevent circular imports
//------------------------------------------------------

export const RoomSelectSchema: z.ZodType<Prisma.RoomSelect> = z.object({
  id: z.boolean().optional(),
  name: z.boolean().optional(),
  createdAt: z.boolean().optional(),
  updatedAt: z.boolean().optional(),
  furniture: z.union([z.boolean(),z.lazy(() => FurnitureFindManyArgsSchema)]).optional(),
  companions: z.union([z.boolean(),z.lazy(() => CompanionFindManyArgsSchema)]).optional(),
  _count: z.union([z.boolean(),z.lazy(() => RoomCountOutputTypeArgsSchema)]).optional(),
}).strict()

export const RoomUpsertArgsSchema: z.ZodType<Prisma.RoomUpsertArgs> = z.object({
  select: RoomSelectSchema.optional(),
  include: z.lazy(() => RoomIncludeSchema).optional(),
  where: RoomWhereUniqueInputSchema,
  create: z.union([ RoomCreateInputSchema,RoomUncheckedCreateInputSchema ]),
  update: z.union([ RoomUpdateInputSchema,RoomUncheckedUpdateInputSchema ]),
}).strict() as z.ZodType<Prisma.RoomUpsertArgs>;

export default RoomUpsertArgsSchema;
